Split out action into separate file
